+++ Packages/com.texelsaur.common/Runtime/Scripts/Base/BasicTest.cs
@@ -1,0 +1,13 @@
+﻿
+using UdonSharp;
+using UnityEngine;
+using VRC.SDKBase;
+using VRC.Udon;
+
+namespace Texel
+{
+    public abstract class BasicTest : UdonSharpBehaviour
+    {
+        public abstract bool _Test();
+    }
+}